Extract actions from widget component
